+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Add `Falcon::Limiter::LongTask#pending?` for detecting delayed long tasks which have not acquired yet.
+
 ## v0.3.0
 
   - Use `Async::Limiter::Token#close` when closing sockets so cached tokens cannot re-acquire after socket close.